#67 restructure some code
