--- src/main/java/de/yfu/intranet/methodendb/configs/WebSecurityConfig.java
@@ -1,51 +1,0 @@
-package de.yfu.intranet.methodendb.configs;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpMethod;
-import org.springframework.security.config.annotation.web.WebSecurityConfigurer;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.CorsConfigurationSource;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-
-import java.util.Arrays;
-
-@Configuration
-public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
-
-
-    private final CorsConfig corsConfig;
-
-    public WebSecurityConfig(final CorsConfig corsConfig) {
-        this.corsConfig = corsConfig;
-    }
-
-    @Bean
-    public CorsConfigurationSource corsConfigurationSource() {
-        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        source.registerCorsConfiguration("/**", corsConfig);
-        System.out.println("CORS Config: " + corsConfig.getAllowedOrigins());
-        return source;
-    }
-
-
-    public void configure(HttpSecurity http) throws Exception {
-        http
-            .cors()
-            .and()
-            .csrf().disable()
-            .authorizeRequests()
-                .antMatchers(HttpMethod.OPTIONS,"/*", "/**/*", "/")
-                .permitAll();
-    }
-
-    @Configuration
-    @ConfigurationProperties(prefix = "cors")
-    public static class CorsConfig extends CorsConfiguration {}
-
-}
